requirements.txt
numpy
torch
torchvision
matplotlib
tqdm
Pillow
scikit-image
transformations
pytorch3d